+++ lib/gamtools/data/include/slice.h
